refactor: timeline.module.css→styleの名前がわかりずらかったので直しました。timeline.tsxもそれに合わせて修正は言ってます。Container内のheightいらなかったので消しました。 menu.tsx→フォルダの位置をUIに移動しました。その関係でcssも移動してるのと使ってるコードは更新かかってます。 appLayout.module.css→overflowがautoだったのでhiddenに変更しました。 feature/#112
