+++ frontend/src/components/layouts/appLayout/appLayout.module.css
@@ -17,7 +17,7 @@
 }
 
 .appLayoutRight {
-  overflow-y: auto;
+  overflow: hidden;
 }
 
 @media (max-width: 960px) {
