+++ frontend/src/features/timeline/styles/timeline.module.css
@@ -12,11 +12,11 @@
   grid-template-columns: 1fr 40%;
 }
 
-.timelineContainerLeft {
+.timelineMain {
   overflow-y: auto;
 }
 
-.timelineContainerRight {
+.timelineInfoBox {
   border-left: 1px solid rgb(50, 85, 150);
   overflow: auto;
   min-width: 0;
@@ -30,16 +30,15 @@
   .timelineContainer {
     display: flex;
     flex-direction: column;
-    height: 100%;
   }
-  .timelineContainerLeft {
+  .timelineMain {
     order: 2;
     width: 100%;
     display: flex;
     flex-direction: column;
     flex: 1;
   }
-  .timelineContainerRight {
+  .timelineInfoBox {
     display: none;
   }
   .modalButton {
